requirements.txt
langchain==0.1.11
langchain-community==0.0.27
langchain-openai==0.0.8
langchain-voyageai==0.1.1
faiss-cpu==1.8.0
openai==1.12.0
python-dotenv==1.0.0